+++ src/main/java/com/idark/valoria/registries/item/types/elemental/AquariusScytheItem.java
@@ -22,6 +22,8 @@ import pro.komaru.tridot.util.math.*;
 
 import java.util.*;
 
+import static com.idark.valoria.Valoria.BASE_ATTACK_RADIUS_UUID;
+
 public class AquariusScytheItem extends ScytheItem{
     public ArcRandom arcRandom = Tmp.rnd;
     private final float attackDamage;
@@ -34,6 +36,7 @@ public class AquariusScytheItem extends ScytheItem{
         builder.put(AttributeReg.DEPTH_DAMAGE.get(), new AttributeModifier(Valoria.BASE_DEPTH_DAMAGE_UUID, "Weapon modifier", 2, AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Weapon modifier", this.attackDamage - 2, AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Weapon modifier", attackSpeedIn, AttributeModifier.Operation.ADDITION));
+        builder.put(AttributeReg.ATTACK_RADIUS.get(), new AttributeModifier(BASE_ATTACK_RADIUS_UUID, "Tool modifier", 3, AttributeModifier.Operation.ADDITION));
         this.defaultModifiers = builder.build();
     }