+++ src/main/java/com/idark/valoria/registries/item/types/elemental/NatureScytheItem.java
@@ -8,6 +8,8 @@ import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.ai.attributes.*;
 import net.minecraft.world.item.*;
 
+import static com.idark.valoria.Valoria.BASE_ATTACK_RADIUS_UUID;
+
 public class NatureScytheItem extends ScytheItem{
     private final float attackDamage;
     private final Multimap<Attribute, AttributeModifier> defaultModifiers;
@@ -18,6 +20,7 @@ public class NatureScytheItem extends ScytheItem{
         builder.put(AttributeReg.NATURE_DAMAGE.get(), new AttributeModifier(Valoria.BASE_NATURE_DAMAGE_UUID, "Weapon modifier", 2, AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Weapon modifier", this.attackDamage - 2, AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Weapon modifier", attackSpeedIn, AttributeModifier.Operation.ADDITION));
+        builder.put(AttributeReg.ATTACK_RADIUS.get(), new AttributeModifier(BASE_ATTACK_RADIUS_UUID, "Tool modifier", 3, AttributeModifier.Operation.ADDITION));
         this.defaultModifiers = builder.build();
     }